chore(refactore): added resolvers to core module
