+++ web/src/app/main/main-routing.module.ts
@@ -6,9 +6,9 @@ import { MainComponent } from './main.component';
 import { HomepageComponent } from '../homepage/homepage.component';
 import { CreateProjectComponent } from '../projects/create/create.component';
 import { ViewProjectComponent } from '../projects/view/view.component';
-import { ViewProjectResolver } from '../../resolvers/view-project.resolver';
+import { ViewProjectResolver } from '../core/resolvers/view-project.resolver';
 import { EditProjectComponent } from '../projects/edit/edit.component';
-import { EditProjectResolver } from '../../resolvers/edit-project.resolver';
+import { EditProjectResolver } from '../core/resolvers/edit-project.resolver';
 import { PrivateProjectsComponent } from '../projects/private/private.component';
 import { ProfileComponent } from '../profile/profile.component';
 import { FeaturesComponent } from '../features/features.component';
